Refactored server.ts and related files. Changed procfile to run node instead of ts-node
