+++ packages/commonwealth/Procfile
@@ -1,3 +1,3 @@
-web: cd packages/commonwealth && NODE_OPTIONS=--max-old-space-size=$(../../scripts/get-max-old-space-size.sh) ts-node -P tsconfig.server.json -T server.ts
+web: cd packages/commonwealth && NODE_OPTIONS=--max-old-space-size=$(../../scripts/get-max-old-space-size.sh) node ./build/commonwealth/server.js
 consumer: cd packages/commonwealth && node --max-old-space-size=$(../../scripts/get-max-old-space-size.sh) build/commonwealth/server/CommonwealthConsumer/CommonwealthConsumer.js run-as-script
 release: cd packages/commonwealth && npx sequelize-cli db:migrate --config server/sequelize.json
