+++ packages/commonwealth/server/scripts/setupAppRoutes.ts
@@ -1,14 +1,21 @@
 import cheerio from 'cheerio';
 import { ChainBase, ChainNetwork, ProposalType } from 'common-common/src/types';
-import { DEFAULT_COMMONWEALTH_LOGO } from '../config';
+import express from 'express';
+import fs from 'fs';
+import * as path from 'path';
+import devWebpackConfig from '../../webpack/webpack.dev.config';
+import prodWebpackConfig from '../../webpack/webpack.prod.config';
+import { DEFAULT_COMMONWEALTH_LOGO, DEV } from '../config';
 import type { DB } from '../models';
 import type { ChainInstance } from '../models/chain';
 import { factory, formatFilename } from 'common-common/src/logging';
+import webpack from 'webpack';
+import webpackDevMiddleware from 'webpack-dev-middleware';
+import webpackHotMiddleware from 'webpack-hot-middleware';
 
 const log = factory.getLogger(formatFilename(__filename));
 
 const NO_CLIENT_SERVER = process.env.NO_CLIENT === 'true';
-const DEV = process.env.NODE_ENV !== 'production';
 
 function cleanMalformedUrl(str: string) {
   return str.replace(/.*(https:\/\/.*https:\/\/)/, '$1');
@@ -25,29 +32,31 @@ const decodeTitle = (title: string) => {
 
 const setupAppRoutes = (
   app,
-  models: DB,
-  devMiddleware,
-  templateFile,
-  sendFile
+  models: DB
 ) => {
   if (NO_CLIENT_SERVER) {
     return;
   }
-  log.info('setupAppRoutes');
-  // Development: serve everything through devMiddleware
-  if (DEV) {
-    app.get('*', (req, res, next) => {
-      req.url = '/build/';
-      devMiddleware(req, res, next);
-    });
-    return;
-  }
 
+  log.info('setupAppRoutes');
+
+  // if we are using an external build, serve it from current folder, otherwise server from build folder.
+  const indexHtmlPath = `${__dirname}`.includes('/build/commonwealth') ?
+    path.resolve(`${__dirname}/../../../index.html`) :
+    path.resolve(`${__dirname}/../../build/index.html`)
+
+  const templateFile = (() => {
+    try {
+      return fs.readFileSync(indexHtmlPath);
+    } catch (e) {
+      console.error(`Failed to read template file: ${e.message}`);
+    }
+  })();
   // Production: serve SEO-optimized routes where possible
   //
   // Retrieve the default bundle from /build/index.html, and overwrite <meta>
   // tags with data fetched from the backend.
-  if (!templateFile) {
+  if (process.env.NODE_ENV === 'production' && !templateFile) {
     throw new Error('Template not found, cannot start production server');
   }
 
@@ -225,10 +234,35 @@ const setupAppRoutes = (
     await renderProposal(scope, proposalType, proposalId, res, chain);
   });
 
-  app.get('*', (req, res) => {
-    log.info(`setupAppRoutes sendFiles ${req.path}`);
-    sendFile(res);
-  });
+  // serve the compiled app
+  if (!NO_CLIENT_SERVER) {
+    if (DEV) {
+      const compiler = DEV
+        ? webpack(devWebpackConfig as any)
+        : webpack(prodWebpackConfig as any);
+      const devMiddleware =
+        DEV && !NO_CLIENT_SERVER
+          ? webpackDevMiddleware(compiler as any, {
+            publicPath: '/build',
+          })
+          : null;
+      app.use(devMiddleware);
+      app.use(webpackHotMiddleware(compiler));
+
+      // Development: serve everything through devMiddleware
+      app.get('*', (req, res, next) => {
+        req.url = '/build/';
+        devMiddleware(req, res, next);
+      });
+    } else {
+      app.use('/build', express.static('build'));
+
+      app.get('*', (req, res) => {
+        log.info(`setupAppRoutes sendFiles ${req.path}`);
+        res.sendFile(indexHtmlPath);
+      });
+    }
+  }
 };
 
 export default setupAppRoutes;
