+++ packages/commonwealth/webpack/webpack.base.config.js
@@ -1,6 +1,7 @@
 const path = require('path');
 const webpack = require('webpack');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const HtmlWebpackInjectAttributesPlugin = require('html-webpack-inject-attributes-plugin');
 
 require('dotenv').config();
 
@@ -30,14 +31,13 @@ module.exports = {
         process.env.DISCORD_UI_URL || 'http://localhost:3000'
       ),
     }),
-    new CopyWebpackPlugin({
-      patterns: [
-        {
-          from: path.resolve(__dirname, '../favicon.ico'),
-          to: path.resolve(__dirname, '../build/commonwealth/favicon.ico'),
-        },
-      ],
+    new HtmlWebpackPlugin({
+      template: path.resolve(__dirname, '../client/index.html'),
+      attributes: {
+        'data-cfasync': 'false',
+      },
     }),
+    new HtmlWebpackInjectAttributesPlugin(),
     new webpack.IgnorePlugin({
       resourceRegExp: /^\.\/locale$/,
       contextRegExp: /moment$/,
